requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-allauth==65.6.0
django-apscheduler==0.7.0
django-cors-headers==4.7.0
django-filter==25.1
django-modeltranslation==0.19.13
django-tinymce==4.1.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.8.1
et_xmlfile==2.0.0
Faker==37.5.3
frozenlist==1.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.2
msgpack==1.1.0
multidict==6.6.3
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
prompt_toolkit==3.0.50
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
service-identity==24.2.0
setuptools==80.0.1
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
zope.interface==7.2
